fix Adopt creating methods on the wrong class
Checking for $class->can('has') is not a good way to check if a class
has already had Moo imported into it.  In the case of Adopt, the @ISA
would already be set to Class::Accessor::Fast, which would be a Moo
class and thus have a 'has' sub.  ->can would find this, it would then
be used to add attributes, which would be created on
Class::Accessor::Fast rather than the intended subclass.

The next version of Moo has extra checks for modifying a constructor
after it has been used, which detected this problem.